media-source/SourceBuffer-remove_readyState.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>SourceBuffer#remove(start, end) with "ended" readyState</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = '/media/white.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
async_test(function(t) {
    GET(FILE, function(data) {
        var mediaSource = new MediaSource();
        var num_sourceopen = 0;
        var events = [];
        mediaSource.addEventListener('sourceopen', function(e){
            events.push('sourceopen');
            num_sourceopen++;
            if(num_sourceopen > 1) {
                mediaSource.endOfStream();
                return;
            }
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            sourceBuffer.addEventListener('updateend', function(e) {
                if(mediaSource.readyState != 'ended') {
                    mediaSource.endOfStream();//make mediaSource.readyState 'ended'

                    sourceBuffer.remove(1,2);
                    t.step(function() {
                        assert_equals(mediaSource.readyState, 'open');
                    });
                }
            });
            sourceBuffer.appendBuffer(data);
        });
        mediaSource.addEventListener('sourceended', function(e){
            events.push('sourceended');
            if(num_sourceopen > 1) {
                t.step(function() {
                    assert_array_equals(events,
                                        ['sourceopen','sourceended',
                                         'sourceopen','sourceended']);
                });
                t.done();
            }
        });
        video.src = window.URL.createObjectURL(mediaSource);
    });
},
'SourceBuffer#remove(start, end) :  If the readyState attribute of the parent media source is in the "ended" state then run the following steps:' +
'1. Set the readyState attribute of the parent media source to "open"' +
'2. Queue a task to fire a simple event named sourceopen at the parent media source .');
</script>
</body>
</html>
